feat: Apply MVVM pattern
- Move saveDb, loadDb to SearchDbHelper
- Use coroutine to control functions order
